gpg-agent is a bit stupid when it comes to which tty we are on
create a function of ssh that does some cleanup of our gpg environment
before actually ssh'ing anywhere. this should help prevent most of the
issues with gpg-agent launching pinentry in a tty other than the one i
am in, or worse, one that is logged in remotely. in both cases, nothing
really works and it makes me sad.

in fact, this commit makes me sad, but it's the lesser of two sads.